// src/redux/reducers/layout/actions-types.ts
export enum LayoutActionTypes {
  SET_FOOTER_CONTENT = "SET_FOOTER_CONTENT",
  SET_INFO_DESCRIPTION = "SET_INFO_DESCRIPTION",
  SET_INFO_FEES = "SET_INFO_FEES",
  SHOW_BACK_BUTTON = "SHOW_BACK_BUTTON",
  HIDE_BACK_BUTTON = "HIDE_BACK_BUTTON",
  SHOW_INFO = "SHOW_INFO",
  HIDE_INFO = "HIDE_INFO",
  SHOW_FOOTER = "SHOW_FOOTER",
  HIDE_FOOTER = "HIDE_FOOTER",
  HIDE_FOOTER_WITH_INFO_AND_BACK_BUTTON = "HIDE_FOOTER_WITH_INFO_AND_BACK_BUTTON",
}